+++ include/parse.h
@@ -6,7 +6,7 @@
 /*   By: rmiranda <rmiranda@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 16:03:07 by rmiranda          #+#    #+#             */
-/*   Updated: 2023/10/24 00:04:36 by rmiranda         ###   ########.fr       */
+/*   Updated: 2023/10/23 21:16:59 by rmiranda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,7 @@
 
 // Getter
 int		add_next_instruction(char *line, t_map_info *info_ptr);
+int		parse(t_map_info *info_ptr, char *map_path);
 
 // Validation
 int		map_validation(t_map_info *info_ptr);
@@ -31,5 +32,6 @@ int		add_line_to_map(t_map_info *info_ptr, char *line);
 char	**normalize_map(t_map_info *info_ptr);
 int		assert_is_not_surrounded_by_wall(t_map_info info);
 void	space_injest_func(unsigned int nb, char *c);
+void	parse_destroy(t_map_info info);
 
 #endif
